major refactor (maputils, views): return map markers in groups
JSON-rendering views return doculects in one list.
This is not optimal:

* grouping of doculects is difficult
(e.g. for upcoming interactive list)

* Icon data is repeated constantly in many doculects

Create a list of groups of markers instead, each group
sharing data for the map icon.

If there is only one group of markers, return
a list anyway (consisting of one group of markers)