refactor(`baseDoculectMapAndList.js`): extract DoculectMap, add context
* use React.createContext to create a context
that will be used across the map, the interactive list
and (in query wizard) the form.

* note that a `useState()` hook is needed to set the
value in the context, so pass not the data itself
as `value` of context, but the constants created
by the useState() hook.

* for now, only factor out DoculectMap
to a separate module, plan to add list and form later

* the map fetches no data itself, uses `useContext()`
instead