refactor: move list/legend switch to parent `page_with_map.jinja2`
move <div> with the switch and <script> import into
the template page_with_map.jinja2 that is parent
for feature.jinja2 and families.jinja2.

note that the JS script has now to ensure it only
runs after full rendering of page. Otherwise
it will not wait for the child template
(feature.jinja2 or families.jinja2) to render.