fix `DoculectMap.js`, `MapAndList.js`: remove `console.log`s
the problem with map freezing without markers on it
 seems to be fixed by commit 75b860c4d4c7880faab1d94c9dd857518b3eed18.
(note the typo in commit message: I introduced
one more useState, not useEffect)
The problem seems to have been that
mapRef.current was already not null but the map
was still not fully loaded, so the groups could not
be really added to it.
Adding another useState hook and adding this hook
to the dependency array of useEffect() for rendering
groups seems to have done the trick.