feat(query wizard CSS) make option's bottom-margin dynamic
before I had `margin-bottom` for
`.toggle:checked+.lbl-toggle+.collapsible-content`,
which meant that when a category is expanded,
it always has a big bottom margin for all values
in the last feature of the category to fit in.
This could confuse users.

Instead, add margin only when the drop-down list
is open. That will lead to the list of options itself
being way below the field that was clicked, so it
has to get an equivalent negative top margin.

This will let me accomplish the following:
* the bottom margin is not always there, and
* when it is there, its background is not white,
but of the color of the category, which makes it clear
for the user that there might be another category below.