fix(`DoculectMap.js`) unfocus doculect on `mouseout` in interactive list
addresses #2. When user hovers out of an item
in interactive list, the map must go back to focusing
on the doculect that was indicated in the URL.
This can be achieved with the existing React context:
when user moves the mouse out of the list item,
the `idOfDoculectToOpenPopupOnMap` becomes null.